% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faker.R
\name{faker}
\alias{faker}
\title{faker}
\usage{
faker(resource, ..., quantity = 1, locale = "en_US", seed = NULL)
}
\arguments{
\item{resource}{resorce to get from api}

\item{...}{other params}

\item{quantity}{quantity of resrources to return}

\item{locale}{locale param for api}

\item{seed}{seed param for api}
}
\value{
list of status, api code, total num of resources, data
}
\description{
faker
}
\examples{
str(faker("images", width = 300))
}
