% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faker_person.R
\name{faker_person}
\alias{faker_person}
\title{faker_person}
\usage{
faker_person(
  gender = NULL,
  birthday_start = NULL,
  birthday_end = NULL,
  quantity = 1,
  locale = "en_US",
  seed = NULL
)
}
\arguments{
\item{gender}{string of gender}

\item{birthday_start}{api param}

\item{birthday_end}{api param}

\item{quantity}{api param}

\item{locale}{api param}

\item{seed}{api param}
}
\value{
list of status, api code, total number resrouces returned, data
}
\description{
faker_person
}
\examples{
str(faker_person("male"))
}
